fix(ci): python-coverage-shards (8) — split oversized pre-push finalize post-commit test file under maintainability line limit
test_pr_monitor_pre_push_validation_finalize_post_commit.py was 1746 lines,
exceeding the 1500-line MAX_FIRST_PARTY_FILE_LINES guard enforced by
test_core_decomposition_maintainability. Moved the post-commit committed-delta
ownership re-validation tests (working-tree-only unowned dirt, rename/non-ASCII
dirt, untracked fail-closed, no-commit-clean self-commit re-validation) into a
new test_pr_monitor_pre_push_validation_finalize_post_commit_delta.py module,
following the existing _post_commit_edges.py split convention.
